requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
spotipy==2.19.0
tabulate==0.8.9
urllib3==1.26.8
